+++ dynamite/config_reader.py
@@ -265,16 +265,8 @@ class Configuration(object):
 
                     logger.debug(f"{comp}... instantiating {data_comp['type']} "
                               "object")
-                    if 'contributes_to_potential' not in data_comp:
-                        text = f'Component {comp} needs ' + \
-                                'contributes_to_potential attribute'
-                        logger.error(text)
-                        raise ValueError(text)
-#                    c = globals()[data_comp['type']](contributes_to_potential
-#                        = data_comp['contributes_to_potential'])
-                    c = getattr(physys,data_comp['type'])(name = comp,
-                            contributes_to_potential \
-                            = data_comp['contributes_to_potential'])
+#                    c = globals()[data_comp['type']]()
+                    c = getattr(physys,data_comp['type'])(name = comp)
                     # check for extra config entries/typos before any more
                     # information is read from config file
                     keys_ok = ['parameters', 'type', 'include',
@@ -290,6 +282,12 @@ class Configuration(object):
                             f'Allowed entries: {keys_ok}. Check for typos.'
                         self.logger.error(text)
                         raise ValueError(text)
+                    if 'contributes_to_potential' in data_comp:
+                        text = f'Component {comp}: the attribute ' + \
+                                'contributes_to_potential is DEPRECATED ' + \
+                                'and will be ignored. In a future ' + \
+                                'DYNAMITE release this will cause an error.'
+                        logger.warning(text)
 
                     # initialize the component's parameters, kinematics,
                     # and populations
@@ -323,7 +321,7 @@ class Configuration(object):
                                 logger.warning(f'Kinematics {kin}: the '
                                     '\'weight\' attribute is DEPRECATED and '
                                     'will be ignored. In a future DYNAMITE '
-                                    'realease this will report an error.')
+                                    'release this will cause an error.')
                                 del data_kin['weight']
                             kinematics_set = getattr(kinem,data_kin['type'])\
                                                 (name=kin,